[GraphQL/TransactionBlock] Unchanged Shared Objects
Adds `TransactionBlock.unchangedSharedObjects` which details the
shared objects that a transaction read but didn't modify (because it
only wanted to read it or because it was deleted before the
transaction had a chance to run).

The shape of an `UnchangedSharedObject` corresponding to a read has
been factored into the `ObjectRead` type which is shared with
`ImmutableOrOwned` or `Receiving` transaction inputs.

Test Plan:

New E2E test:

```
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- shared.move
```