[GraphQL/Query] Standalone type query
Makes it possible to get layout information for a type without it
being associated to an existing object, as requested by @hayes-mysten.

Note that currently this implementation will report an internal error
if you try and fetch the layout of a type that doesn't exist. This
should really be a user error (bad user input), for supplying a type
that doesn't exist.

This will be fixed in a follow-up when we overhaul the error story for
GraphQL (Similar fixes are also required for the database and
execution layers as well).

Test Plan:

New E2E tests:

```
sui-graphql-e2e-tests$ cargo nextest run \
  -j 1 --features pg_integration         \
  -- types.move
```